/adhoc: Refactoring configurator/GetConfigFileOrFlag (#203)
* Added file configuration

* Added flag configuration

* Added LoadSpecificConfig

* Refactored configuration_Test

* Matched error handling

* Switched commands to using LoadSpecificConfig

* Clean up commands now using LoadSpecificConfig

* Removed configurator and tests

* Comment

* Added TestFlagWithBadJson

* Changed LoadSpecificConfig to LoadCommandsConfig and updated its' unit tests

* Updated use of LoadSpecificConfig to LoadCommandsConfig

* Update pkg/config/configuration.go


* Update pkg/config/flags.go


* Update pkg/config/file.go


* Finished renaming functions.

* fixes after merge from master

* getConfigName and associated functions moved out of config and into command package

* Moved getConfigName into shore

* Making codacy happy

* trying to make codacy even happier

* Still trying to make codacy happy

* This will surely appease the codacy gods
